+++ NEWS.md
@@ -1,4 +1,5 @@
 # geogenr 2.0.1
+* Data export function in `geomultistar` format: `as_geomultistar()`.
 * Modify the cph field of the description.
 * Include logo on website.